server/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34
pydantic-settings>=2.7
python-dotenv>=1.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.14
redis>=5.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt==4.0.1
email-validator>=2.0
httpx>=0.28
pytest>=8.0
